refactor(persistence): move configuration classes from Infrastructure to Persistence layer
- Renamed configuration files from AirlineBookingSystem.Infrastructure/Configurations to AirlineBookingSystem.Persistence/Configurations
- Updated namespaces from AirlineBookingSystem.Infrastructure.Configurations to AirlineBookingSystem.Persistence.Configurations
- Ensured all entity configuration classes reflect the new layer structure

This refactoring improves project structure by aligning configuration classes with the Persistence layer. No functional changes were introduced.
