+++ AirlineBookingSystem.Persistence/Configurations/PersonConfig.cs
@@ -2,7 +2,7 @@
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
-namespace AirlineBookingSystem.Infrastructure.Configurations
+namespace AirlineBookingSystem.Persistence.Configurations
 {
     public class PersonConfig : IEntityTypeConfiguration<Person>
     {
